Script: fields API for x-pack constant_keyword
Adds the fields API for the constant_keyword field mapper

Moves implementation to `AbstractKeywordDocValuesField`, allowing
code sharing with `KeywordDocValuesField`.

API:
```
field('const').get('default')
field('const').get('default', 0)
```

Refs: #79105
