+++ x-pack/plugin/mapper-constant-keyword/src/main/java/org/elasticsearch/xpack/constantkeyword/ConstantKeywordDocValuesField.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+package org.elasticsearch.xpack.constantkeyword;
+
+import org.elasticsearch.index.fielddata.SortedBinaryDocValues;
+import org.elasticsearch.script.field.AbstractKeywordDocValuesField;
+
+public class ConstantKeywordDocValuesField extends AbstractKeywordDocValuesField {
+    public ConstantKeywordDocValuesField(SortedBinaryDocValues input, String name) {
+        super(input, name);
+    }
+}
